Clean up FestivalTask and fix issue with festival not returning
Make sure to run festival in batch mode (as opposed to interactive
mode) so it returns. Additionally, add variouchecks to deal with
festival hanging, the task being cancelled, or things otherwise not
working correctly.

Temporarily disables the use of SpeechManager to test FestivalTask
properly. This will be restoreonce SpeechManager is s properly
refactored.