Allow for variable speed text-to-speech playback
Using festival scheme commands, allow for thplayback duration
to be altered on a per-task basis. If a speed of zero is given,
the phrase is not played and the task returns a 'failure' status
(this behaviour is built-in to festival).

Adds a Spinner control to the main window to test this.